pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opensecagent"
version = "0.2.0"
description = "Autonomous Server Cybersecurity Expert Bot - monitors hosts and containers, detects risks, recommends and executes safe remediation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "OpenSecAgent Contributors"}]
keywords = ["security", "cybersecurity", "monitoring", "docker", "incident-response"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.0",
    "docker>=7.0.0",
    "aiosmtplib>=3.0.0",
    "openai>=1.12.0",
    "httpx>=0.26.0",
    "filelock>=3.13.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.scripts]
opensecagent = "opensecagent.main:main"
opensecagent-cli = "opensecagent.cli:main"

[project.urls]
Homepage = "https://github.com/opensecagent/opensecagent"
Repository = "https://github.com/opensecagent/opensecagent"
Issues = "https://github.com/opensecagent/opensecagent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["opensecagent*"]

[tool.setuptools.package-data]
opensecagent = ["config/*.yaml"]

